feat: Remove 'lai_vinculadas' field and update latitude field properties
This change removes the 'lai_vinculadas' field from the FIELDS dictionary and updates the properties of the 'latitude_coordenadas' and 'latitude_da_estacao' fields:

- Removes the 'lai_vinculadas' field (field ID 481) from the FIELDS dictionary.
- Updates the 'latitude_coordenadas' field to remove the '_dtype' parameter.
- Updates the 'latitude_da_estacao' field to remove the '_dtype' parameter.

These changes simplify the field configuration and ensure consistent handling of the latitude-related fields.

+++ fiscaliza/attrs.py
@@ -155,14 +155,10 @@ FIELDS = {
         ],
     ),
     # "lai_vinculadas": SimpleField(481, "LAI vinculadas"),
-    "latitude_coordenadas": SimpleField(170, "Latitude (º):", True, _dtype="float"),
-    "latitude_da_estacao": SimpleField(
-        191, "Latitude da estação:", True, _dtype="float"
-    ),
-    "longitude_coordenadas": SimpleField(171, "Longitude (º):", True, _dtype="float"),
-    "longitude_da_estacao": SimpleField(
-        192, "Longitude da estação:", True, _dtype="float"
-    ),
+    "latitude_coordenadas": SimpleField(170, "Latitude (º):", True),
+    "latitude_da_estacao": SimpleField(191, "Latitude da estação:", True),
+    "longitude_coordenadas": SimpleField(171, "Longitude (º):", True),
+    "longitude_da_estacao": SimpleField(192, "Longitude da estação:", True),
     "motivo_de_lai": FieldWithOptions(
         164,
         "Motivo de LAI:",
